Highlight changes in technical skills and employment history in resume page.
In the "Resume" page, two changes were made for accuracy and relevancy of information. In the listed technologies, 'Testing Library' was removed from the 'Frontend' category and added to the 'TDD & BDD' category to better reflect the skill set. Additionally, a grammatical error in the description of the job responsibilities was corrected by changing 'are' to 'were'.